Skip conn e2e on Win. Divide map e2e tests.
